+++ P3 Django/bibliotheca_mobilis/community/forms.py
@@ -1,6 +1,7 @@
 from django import forms
-from community.models import armyModel
-from community.models import charactersModel, armyModel, speciality
+from community.models import armyModel, charactersModel, armyModel, speciality , creationRace
+
+ 
 
 
 class characterForm(forms.ModelForm):
@@ -23,12 +24,14 @@ class characterForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
+        self.fields['race'].queryset = creationRace.objects.none()
+
         self.fields['specialities'].widget.attrs.update(
             {'class': 'formClass specialityForm',
             'onclick' : 'checkCount(this.id);'}
         )
         self.fields['side'].widget.attrs.update(
-            {'onchange' : 'updateRace(this.value);'}
+            {'onchange' : 'updateRaces(this.value);'}
         )
 
         self.fields['biography'].widget.attrs.update(
@@ -38,6 +41,15 @@ class characterForm(forms.ModelForm):
             self.fields[field].widget.attrs.update({
                 'class': 'formClass',
             })
+        
+        if 'side' in self.data:
+            try:
+                sideId = int(self.data.get('side'))
+                self.fields['race'].queryset = creationRace.objects.filter(side_id=sideId).order_by('name')
+            except (ValueError, TypeError):
+                pass  # invalid input from the client; ignore and fallback to empty race queryset
+        elif self.instance.pk:
+            self.fields['race'].queryset = self.instance.side.race_set.order_by('name')
