+++ P3 Django/bibliotheca_mobilis/community/models/__init__.py
@@ -1,0 +1,3 @@
+from .models import *
+from .characterModels import *
+from .armyModels import *
