+++ P3 Django/bibliotheca_mobilis/community/static/community/community.js
@@ -18,4 +18,32 @@ function checkCount(id) {
         alert('Seules trois spécialités peuvent être séléctionnées')
         document.getElementById(id).checked = false
     }
+
+}    
+
+function updateRaces(sideId) {
+    var xhttp = new XMLHttpRequest();
+    var url= document.getElementById('characterCreationForm').getAttribute('data-races-url');
+    //  add parameter to GET URL 
+    url+= '?sideId=' + sideId
+
+    // function called when server answer arrived 
+    xhttp.onreadystatechange = function() {
+        // readyState == 4 (done) the request sent and answer received
+        // status == 200 the server answer all is ok
+        if (this.readyState == 4 && this.status == 200) {
+
+            // update race list against side selection
+            document.getElementById('id_race').innerHTML = this.responseText;
+            
+            // remove null option after first select
+            var nullOption = document.getElementById('id_side').getElementsByTagName('option')[0];
+            if (nullOption.value == '' ) {
+                nullOption.remove();
+            };
+        };
+    }
+    xhttp.open("GET", url, true);
+    xhttp.send();   
 }
+
