+++ P3 Django/bibliotheca_mobilis/community/templates/community/characterCreation.html
@@ -16,10 +16,10 @@
 <div class='textPosition'>
     {% if user.is_authenticated%}
 
-    <form action="" method='post' class='formular' enctype="multipart/form-data">
+    <form id="characterCreationForm" action="" method='post' class='formular' enctype="multipart/form-data"
+        data-races-url="{% url 'ajaxLoadRaces' %}" novalidate>
         {% csrf_token %}
 
-        {% comment %} {{form.as_p}} {% endcomment %}
 
         {%for field in form%}
         {% comment %} field.name = nom label (pseudo,email etc) {% endcomment %}
