+++ P3 Django/bibliotheca_mobilis/community/templates/community/race_dropdown_list_options.html
@@ -1,0 +1,3 @@
+{% for race in races %}
+<option value="{{ race.pk }}">{{ race.name }}</option>
+{% endfor %}
