+++ P3 Django/bibliotheca_mobilis/community/views/__init__.py
@@ -1,0 +1,3 @@
+from .armyViews import *
+from .characterViews import *
+from .loadRacesView import *
